Add a quick way to update the spell file
The spell file is always modified, and I'm sick of committing changes to
it manually all the time. This commit adds a script that will commit the
spell file so I don't have to.

Apparently I didn't have GNU grep installed, so that is added in this
commit as well. Plus a bonus little bug fix in the sync script.